requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
attrs==22.2.0
backports.entry-points-selectable==1.1.0
bcrypt==4.0.1
bleach==6.0.0
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.2
Django==3.2.5
docutils==0.19
filelock==3.0.12
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-expects-json==1.7.0
flask-http-middleware==0.2.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
mdurl==0.1.2
more-itertools==9.1.0
mysqlclient==2.1.1
packaging==23.0
pkginfo==1.9.6
platformdirs==2.0.2
Pygments==2.14.0
PyJWT==2.6.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.2
six==1.16.0
SQLAlchemy==2.0.4
sqlparse==0.4.1
tomli==2.0.1
twine==4.0.2
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.6.0
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.11.0
